Move ConditionalDecorator into codegen-client
